airflow/requirements.txt
apache-airflow-providers-google
pandas
chardet
pyarrow
pyspark==3.2.0